chore(i18n): update frontend locales structure
