--- app/javascript/locales.js
@@ -1,91 +1,0 @@
-const messages = {
-  en: {
-    message: {
-      settings: {
-        button: 'Go to dashboard',
-        adminUsers: 'Manage Users',
-        adminRepos: 'Manage Repositories',
-        repoUpdate: 'Last update on',
-        sync: 'sync',
-        loading: '...loading',
-        error: 'There was an error',
-        disablePublic: 'Disable public dashboard',
-        enablePublic: 'Enable public dashboard',
-      },
-      profile: {
-        noTeams: 'No teams',
-        teamsDropdownTitle: 'Teams',
-        noOrganizations: 'No organizations',
-        organizationsDropdownTitle: 'Organizations',
-        recommendedReviewers: 'You should assign your next PR to...',
-        notRecommendedReviewers: 'Avoid assigning your next PR to...',
-        statistics: {
-          obedientCount: 'Obeyed recommendations',
-          indifferentCount: 'Ignored recommendations',
-          rebelCount: 'Rejected recommendations',
-        },
-        relationsTitle: 'Your relation with the rest of the team: ',
-        badgeExplainer: {
-          low: 'Also part of your team! Don\'t ignore them :(',
-          midlow: 'Could still receive some love!',
-          good: 'Your Froggo objective',
-          midhigh: 'Don\'t overdo it...',
-          high: 'You seem fixated...',
-        },
-      },
-      admin: {
-        noDefaultTeam: 'No default team',
-        defaultTeamDropdownTitle: 'Default team',
-      },
-    },
-  },
-  es: {
-    message: {
-      settings: {
-        button: 'Ir a dashboard',
-        adminUsers: 'Administrar Usuarios',
-        adminRepos: 'Administrar Repositorios',
-        repoUpdate: 'Ultima vez actualizado el',
-        sync: 'actualizar',
-        loading: '...cargando',
-        error: 'Ocurrió un error',
-        disablePublic: 'Deshabilitar dashboard público',
-        enablePublic: 'Habilitar dashboard público',
-      },
-      profile: {
-        noTeams: 'Sin equipos',
-        teamsDropdownTitle: 'Equipos',
-        noOrganizations: 'Sin organizaciones',
-        organizationsDropdownTitle: 'Organizaciones',
-        recommendedReviewers: 'Deberías mandar tu próximo PR a...',
-        notRecommendedReviewers: 'Evita mandar tu próximo PR a...',
-        statistics: {
-          obedientCount: 'Recomendaciones seguidas',
-          indifferentCount: 'Recomendaciones ignoradas',
-          rebelCount: 'Recomendaciones no seguidas',
-        },
-        relationsTitle: 'Tu relación con el resto del equipo: ',
-        badgeExplainer: {
-          low: '¡También es parte de tu equipo! No lo ignores :(',
-          midlow: '¡Todavía puedo ayudarte más!',
-          good: 'Tu objetivo Froggo',
-          midhigh: 'No te pases...',
-          high: 'Parece que tienes una fijación...',
-        },
-      },
-      admin: {
-        noDefaultTeam: 'Sin equipo default',
-        defaultTeamDropdownTitle: 'Equipo default',
-      }
-    },
-  },
-  ja: {
-    message: {
-      settings: {
-        button: 'こんにちは、世界',
-      },
-    },
-  },
-};
-
-export default { messages };
