+++ app/javascript/locales/en.js
@@ -1,0 +1,41 @@
+export default {
+  message: {
+    settings: {
+      button: 'Go to dashboard',
+      adminUsers: 'Manage Users',
+      adminRepos: 'Manage Repositories',
+      repoUpdate: 'Last update on',
+      sync: 'sync',
+      loading: '...loading',
+      error: 'There was an error',
+      disablePublic: 'Disable public dashboard',
+      enablePublic: 'Enable public dashboard',
+    },
+    profile: {
+      noTeams: 'No teams',
+      teamsDropdownTitle: 'Teams',
+      noOrganizations: 'No organizations',
+      organizationsDropdownTitle: 'Organizations',
+      recommendedReviewers: 'You should assign your next PR to...',
+      notRecommendedReviewers: 'Avoid assigning your next PR to...',
+      statistics: {
+        obedientCount: 'Obeyed recommendations',
+        indifferentCount: 'Ignored recommendations',
+        rebelCount: 'Rejected recommendations',
+      },
+      timespanDropdownTitle: 'Since',
+      relationsTitle: 'Your relation with the rest of the team: ',
+      badgeExplainer: {
+        low: 'Also part of your team! Don\'t ignore them :(',
+        midlow: 'Could still receive some love!',
+        good: 'Your Froggo objective',
+        midhigh: 'Don\'t overdo it...',
+        high: 'You seem fixated...',
+      },
+    },
+    admin: {
+      noDefaultTeam: 'No default team',
+      defaultTeamDropdownTitle: 'Default team',
+    },
+  },
+};
