+++ app/javascript/locales/es.js
@@ -1,0 +1,41 @@
+export default {
+  message: {
+    settings: {
+      button: 'Ir a dashboard',
+      adminUsers: 'Administrar Usuarios',
+      adminRepos: 'Administrar Repositorios',
+      repoUpdate: 'Ultima vez actualizado el',
+      sync: 'actualizar',
+      loading: '...cargando',
+      error: 'Ocurrió un error',
+      disablePublic: 'Deshabilitar dashboard público',
+      enablePublic: 'Habilitar dashboard público',
+    },
+    profile: {
+      noTeams: 'Sin equipos',
+      teamsDropdownTitle: 'Equipos',
+      noOrganizations: 'Sin organizaciones',
+      organizationsDropdownTitle: 'Organizaciones',
+      recommendedReviewers: 'Deberías mandar tu próximo PR a...',
+      notRecommendedReviewers: 'Evita mandar tu próximo PR a...',
+      statistics: {
+        obedientCount: 'Recomendaciones seguidas',
+        indifferentCount: 'Recomendaciones ignoradas',
+        rebelCount: 'Recomendaciones no seguidas',
+      },
+      timespanDropdownTitle: 'Desde hace',
+      relationsTitle: 'Tu relación con el resto del equipo: ',
+      badgeExplainer: {
+        low: '¡También es parte de tu equipo! No lo ignores :(',
+        midlow: '¡Todavía puedo ayudarte más!',
+        good: 'Tu objetivo Froggo',
+        midhigh: 'No te pases...',
+        high: 'Parece que tienes una fijación...',
+      },
+    },
+    admin: {
+      noDefaultTeam: 'Sin equipo default',
+      defaultTeamDropdownTitle: 'Equipo default',
+    },
+  },
+};
