+++ app/javascript/locales/locales.js
@@ -1,0 +1,9 @@
+import enLocale from './en';
+import esLocale from './es';
+
+const messages = {
+  en: enLocale,
+  es: esLocale,
+};
+
+export default { messages };
